requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.10.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.2.0
requests>=2.28.0
pytrends>=4.9.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.20.0